# docs/openapi/audit.v1.yaml
openapi: 3.1.0
info:
  title: Fact0 Audit API
  version: 1.0.0
  description: |
    Self-hosted audit records with a per-workspace SHA-256 hash chain.
    Verification checks recorded data integrity; it does not establish truth or completeness.

    Authentication uses a write/read API key in `Authorization: Bearer f0_live_*`.
    OSS ingestion is synchronous: single events return 201 and batches return 200.
    Requests are limited to 4 MiB of encoded JSON; oversized requests return 413.
    A batch can contain per-item rejections: always inspect accepted/rejected/errors.
    Hashes use `sha256:v2:` canonical JSON, exact decimal numbers, all immutable
    public event fields, and UTC timestamps normalized to microseconds.
servers:
  - url: http://localhost:3000
    description: Self-hosted web and API origin
security:
  - apiKey: []
tags:
  - name: Ingest
  - name: Query
  - name: Verify
  - name: Export
  - name: Stream
paths:
  /v1/events:
    post:
      tags: [Ingest]
      summary: Ingest a single audit event
      operationId: logEvent
      security:
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/SyncHeader'
        - $ref: '#/components/parameters/SyncQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditEventInput'
      responses:
        '201':
          description: Event committed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '413':
          description: Encoded request exceeds the 4 MiB limit; split the batch or report the individual item as incomplete.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Query]
      summary: List audit events
      operationId: listEvents
      parameters:
        - name: session_id
          in: query
          description: Claude Code lifecycle resource ID or metadata.session_id
          schema:
            type: string
        - name: actor_id
          in: query
          schema:
            type: string
        - name: actor_type
          in: query
          schema:
            type: string
            enum: [human, agent, system]
        - name: action
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
        - name: outcome
          in: query
          schema:
            type: string
            enum: [success, failure, error]
        - name: from
          in: query
          description: RFC3339 UTC lower bound
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: RFC3339 UTC upper bound
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 1000
      responses:
        '200':
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '413':
          description: Encoded request exceeds the 4 MiB limit; split the batch or report the individual item as incomplete.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/events/batch:
    post:
      tags: [Ingest]
      summary: Ingest up to 1000 audit events
      operationId: logBatch
      parameters:
        - $ref: '#/components/parameters/SyncHeader'
        - $ref: '#/components/parameters/SyncQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [events]
              properties:
                events:
                  type: array
                  maxItems: 1000
                  items:
                    $ref: '#/components/schemas/AuditEventInput'
      responses:
        '200':
          description: Batch committed; inspect rejected/errors even on HTTP 200
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/events/{id}:
    get:
      tags: [Query]
      summary: Get a single audit event
      operationId: getEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEvent'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/events/{id}/verify:
    get:
      tags: [Verify]
      summary: Verify a single event hash
      operationId: verifyEvent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Single-event verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleVerifyResult'
  /v1/verify:
    get:
      tags: [Verify]
      summary: Walk the hash chain and verify integrity
      description: Omitted bounds cover all supported recorded history, from 1970-01-01T00:00:00Z through 9999-12-31T23:59:59.999999999Z, including future-dated captures already received.
      operationId: verifyChain
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: scan_all
          in: query
          description: Return all broken events instead of stopping at the first
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Chain verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResult'
  /v1/export/pdf:
    get:
      tags: [Export]
      summary: Download recorded audit PDF report
      description: Omitted bounds cover the same full supported history as verification, including future-dated captures already received. Supply from or to to restrict capture timestamps.
      operationId: exportPDF
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /v1/export/evidence-pack:
    get:
      tags: [Export]
      summary: Download ZIP evidence pack (PDF + metadata)
      description: Omitted bounds cover the same full supported history as verification, including future-dated captures already received. The PDF and verification JSON use identical bounds.
      operationId: exportEvidencePack
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: ZIP file
          content:
            application/zip:
              schema:
                type: string
                format: binary
  /v1/events/stream:
    get:
      tags: [Stream]
      summary: Live SSE stream of new audit events
      operationId: streamEvents
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: API key prefixed with `alk_live_`. Write scope required for ingest.
    shareToken:
      type: http
      scheme: bearer
      description: Share link bearer token
  parameters:
    SyncHeader:
      name: X-Fact0-Sync
      in: header
      description: Set to `true` for synchronous ingest
      schema:
        type: string
        enum: ['true']
    SyncQuery:
      name: sync
      in: query
      description: Set to `true` for synchronous ingest
      schema:
        type: string
        enum: ['true']
  schemas:
    Actor:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [human, agent, system]
        email:
          type: string
    Resource:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
        name:
          type: string
    AuditEventInput:
      type: object
      required: [actor, action, resource, outcome]
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
          description: Optional capture time; defaults to server time. Explicit values must be between 1970-01-01T00:00:00Z and 9999-12-31T23:59:59.999999999Z inclusive. Future-dated captures within this range are accepted and included in default verification and exports. Storage retains microsecond precision.
        actor:
          $ref: '#/components/schemas/Actor'
        action:
          type: string
          description: Dot-notation action name (e.g. document.delete)
        resource:
          $ref: '#/components/schemas/Resource'
        outcome:
          type: string
          enum: [success, failure, error]
        metadata:
          type: object
          additionalProperties: true
    AuditEvent:
      allOf:
        - $ref: '#/components/schemas/AuditEventInput'
        - type: object
          properties:
            tenant_id:
              type: string
            prev_hash:
              type: string
            hash:
              type: string
              pattern: "^sha256:v2:[0-9a-f]{64}$"
            sequence_number:
              type: integer
              format: int64
    BatchResult:
      type: object
      properties:
        accepted:
          type: integer
        rejected:
          type: integer
        ids:
          type: array
          items:
            type: string
        errors:
          type: array
          items:
            $ref: '#/components/schemas/BatchError'
    BatchError:
      type: object
      properties:
        index:
          type: integer
        code:
          type: string
        message:
          type: string
    AsyncReceipt:
      allOf:
        - $ref: '#/components/schemas/BatchResult'
        - type: object
          properties:
            receipt_id:
              type: string
            status:
              type: string
              enum: [queued, committed, failed]
            error_message:
              type: string
            committed_at:
              type: string
              format: date-time
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    VerifyResult:
      type: object
      properties:
        valid:
          type: boolean
        tenant_id:
          type: string
        events_checked:
          type: integer
        root_hash:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        first_broken_event_id:
          type: string
        reason:
          type: string
        broken_count:
          type: integer
        all_broken_events:
          type: array
          items:
            $ref: '#/components/schemas/BrokenEvent'
    BrokenEvent:
      type: object
      properties:
        event_id:
          type: string
        sequence_number:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        action:
          type: string
        actor_id:
          type: string
        reason:
          type: string
    SingleVerifyResult:
      $ref: "#/components/schemas/VerifyResult"
    ReanchorResponse:
      type: object
      properties:
        status:
          type: string
        meta_event_id:
          type: string
        meta_event_seq:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              properties:
                code:
                  type: string
                message:
                  type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Key lacks required scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Duplicate event ID
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'